+++ app/src/main/java/woowacourse/shopping/productdetail/navigator/ProductDetailNavigator.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.productdetail
+package woowacourse.shopping.productdetail.navigator
 
 import woowacourse.shopping.model.ProductUiModel
