+++ app/src/main/java/woowacourse/shopping/productdetail/navigator/ProductDetailNavigatorImpl.kt
@@ -1,7 +1,8 @@
-package woowacourse.shopping.productdetail
+package woowacourse.shopping.productdetail.navigator
 
 import android.content.Context
 import woowacourse.shopping.model.ProductUiModel
+import woowacourse.shopping.productdetail.ProductDetailActivity
 import woowacourse.shopping.shoppingcart.CartActivity
 
 class ProductDetailNavigatorImpl(private val context: Context) : ProductDetailNavigator {
